requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.12.1
beautifulsoup4==4.14.3
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.6
curl-cffi==0.13.0
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
et-xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
frozendict==2.4.7
fsspec==2025.10.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
multitasking==0.0.12
numpy==2.0.2
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
peewee==4.0.4
pillow==11.3.0
platformdirs==4.4.0
postgrest==2.28.3
propcache==0.4.1
protobuf==6.33.6
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.20.0
pyiceberg==0.10.0
pyjwt==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2026.1.post1
pyyaml==6.0.3
razorpay==2.0.1
realtime==2.28.3
requests==2.32.5
rich==14.3.3
scipy==1.13.1
six==1.17.0
slowapi==0.1.9
email-validator==2.3.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
starlette==0.49.3
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
yfinance==1.2.0
